Ok, try this method.

+++ src/pocketmine/level/generator/normal/biome/extension/OceanBiome.php
@@ -45,8 +45,8 @@ class OceanBiome extends WB
 
 		$this->setElevation(46, 63);
 
-		$this->temperature = 0.5F;
-		$this->rainfall = 0.5F;
+		$this->temperature = (float) 0.5;
+		$this->rainfall = (float) 0.5;
 	}
 
 	public function getName() : string
